requirements.txt
fastapi==0.85.1
dependency-injector==4.39.1
uvicorn==0.18.2
python-multipart==0.0.5
tensorflow-cpu==2.10.0
tensorflow-addons==0.18.0
pydantic[dotenv]==1.10.2
boto3==1.26.16
boto3-stubs[s3]==1.26.16
mypy-boto3==1.26.16
pytest==7.2.0
pytest-env==0.8.1
moto==4.0.10
black==22.10.0
